MF-13 - Add user id
