+++ readers/api/endpoint.go
@@ -6,7 +6,6 @@ package api
 import (
 	"context"
 
-	"github.com/MainfluxLabs/mainflux"
 	"github.com/MainfluxLabs/mainflux/pkg/errors"
 	"github.com/MainfluxLabs/mainflux/readers"
 	"github.com/go-kit/kit/endpoint"
@@ -48,7 +47,7 @@ func listAllMessagesEndpoint(svc readers.MessageRepository) endpoint.Endpoint {
 			return nil, err
 		}
 		// Check if user is authorized to read all messages
-		if err := authorizeUser(ctx, user.id, "authorities", "member"); err != nil {
+		if err := authorizeAdmin(ctx, user.id, "authorities", "member"); err != nil {
 			return nil, errors.Wrap(errors.ErrAuthentication, err)
 		}
 
@@ -68,32 +67,4 @@ func listAllMessagesEndpoint(svc readers.MessageRepository) endpoint.Endpoint {
 type userIdentity struct {
 	id    string
 	email string
-}
-
-func identify(ctx context.Context, token string) (userIdentity, error) {
-	var auth mainflux.AuthServiceClient
-	identity, err := auth.Identify(ctx, &mainflux.Token{Value: token})
-	if err != nil {
-		return userIdentity{}, errors.Wrap(errors.ErrAuthentication, err)
-	}
-
-	return userIdentity{identity.Id, identity.Email}, nil
-}
-
-func authorizeUser(ctx context.Context, subject, object, relation string) error {
-	req := &mainflux.AuthorizeReq{
-		Sub: subject,
-		Obj: object,
-		Act: relation,
-	}
-	var auth mainflux.AuthServiceClient
-	res, err := auth.Authorize(ctx, req)
-	if err != nil {
-		return errors.Wrap(errors.ErrAuthorization, err)
-	}
-	if !res.GetAuthorized() {
-		return errors.ErrAuthorization
-	}
-
-	return nil
 }
